+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - cosmwasm-schema: The schema export now doesn't overwrite existing
   `additionalProperties` values anymore ([#2310])
+- cosmwasm-std: Added new `EurekaMsg` and `CosmosMsg::Eureka` variant ([#2340])
 
 [#2310]: https://github.com/CosmWasm/cosmwasm/pull/2310